+++ wp-content/plugins/geoip-detect/api.php
@@ -19,6 +19,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 */
 
 use YellowTree\GeoipDetect\DataSources\DataSourceRegistry;
+
 /**
  * Get Geo-Information for a specific IP
  * @param string 			$ip 		IP-Adress (IPv4 or IPv6). 'me' is the current IP of the server.
@@ -26,6 +27,7 @@ use YellowTree\GeoipDetect\DataSources\DataSourceRegistry;
  * 										from most preferred to least preferred. (Default: Site language, en)
  * @param array				Property names with options.
  * 		@param boolean 		$skipCache		TRUE: Do not use cache for this request. (Default: FALSE)
+ * 		@param string       $source         Change the source for this request only. (Valid values: 'auto', 'manual', 'precision', 'header', 'hostinfo')
  * 		@param float 		$timeout		Total transaction timeout in seconds (Precision+HostIP.info API only) 
  * 		@param int			$connectTimeout Initial connection timeout in seconds (Precision API only)
  * @return YellowTree\GeoipDetect\DataSources\City	GeoInformation. (Actually, this is a subclass of \GeoIp2\Model\City)
@@ -36,6 +38,7 @@ use YellowTree\GeoipDetect\DataSources\DataSourceRegistry;
  * @since 2.0.0
  * @since 2.4.0 New parameter $skipCache
  * @since 2.5.0 Parameter $skipCache has been renamed to $options with 'skipCache' property
+ * @since 2.7.0 Parameter $options['source'] has been introduced
  */
 function geoip_detect2_get_info_from_ip($ip, $locales = null, $options = array()) {
 	_geoip_maybe_disable_pagecache();
@@ -56,7 +59,7 @@ function geoip_detect2_get_info_from_ip($ip, $locales = null, $options = array()
 	
 	// Have a look at the cache first
 	if (!$options['skipCache']) {
-		$data = _geoip_detect2_get_data_from_cache($ip);
+		$data = _geoip_detect2_get_data_from_cache($ip, $options['source']);
 	}
 	
 	if (!$data) {
@@ -100,12 +103,16 @@ function geoip_detect2_get_info_from_ip($ip, $locales = null, $options = array()
  * @param array(string)		$locales	List of locale codes to use in name property
  * 										from most preferred to least preferred. (Default: Site language, en)
  * @param array				Property names with options.
- * 		@param boolean $skipCache	TRUE: Do not use cache for this request. (Default: FALSE) 
+ * 		@param boolean 		$skipCache		TRUE: Do not use cache for this request. (Default: FALSE)
+ * 		@param string       $source         Change the source for this request only. (Valid values: 'auto', 'manual', 'precision', 'header', 'hostinfo')
+ * 		@param float 		$timeout		Total transaction timeout in seconds (Precision+HostIP.info API only) 
+ * 		@param int			$connectTimeout Initial connection timeout in seconds (Precision API only)
  * @return YellowTree\GeoipDetect\DataSources\City	GeoInformation.
  *
  * @since 2.0.0
  * @since 2.4.0 New parameter $skipCache
  * @since 2.5.0 Parameter $skipCache has been renamed to $options with 'skipCache' property
+ * @since 2.7.0 Parameter $options['source'] has been introduced
  */
 function geoip_detect2_get_info_from_current_ip($locales = null, $options = array()) {
 	return geoip_detect2_get_info_from_ip(geoip_detect2_get_client_ip(), $locales, $options);
@@ -119,15 +126,18 @@ function geoip_detect2_get_info_from_current_ip($locales = null, $options = arra
  * @param array(string)				List of locale codes to use in name property
  * 									from most preferred to least preferred. (Default: Site language, en)
  * @param array				Property names with options.
+ * 		@param string       $source         Change the source for this request only. (Valid values: 'auto', 'manual', 'precision', 'header', 'hostinfo')
  * 		@param float 		$timeout		Total transaction timeout in seconds (Precision+HostIP.info API only) 
- * 		@param float		$connectTimeout Initial connection timeout in seconds (Precision API only)
- * @return \YellowTree\GeoipDetect\DataSources\ReaderInterface 	The reader, ready to do its work. Don't forget to `close()` it afterwards. NULL if file not found (or other problems).
+ * 		@param int			$connectTimeout Initial connection timeout in seconds (Precision API only)
  * 
  * @since 2.0.0
  * @since 2.5.0 new parameter $options
+ * @since 2.7.0 Parameter $options['source'] has been introduced
  */
 function geoip_detect2_get_reader($locales = null, $options = array()) {
 	_geoip_maybe_disable_pagecache();
+	$options = _geoip_detect2_process_options($options);
+
 	return _geoip_detect2_get_reader($locales, false, $sourceIdOut, $options);
 }
 
@@ -143,7 +153,13 @@ function geoip_detect2_get_current_source_description($source = null) {
 	if (is_object($source) && $source instanceof \YellowTree\GeoipDetect\DataSources\City) {
 		$source = $source->extra->source;
 	}
-	$source = DataSourceRegistry::getInstance()->getSource($source);
+	$registry = DataSourceRegistry::getInstance();
+	if (is_null($source)) {
+		$source = $registry->getCurrentSource();
+	} else {
+		$source = $registry->getSource($source);
+	}
+	
 	if ($source) {
 		return $source->getShortLabel();
 	}
@@ -198,7 +214,10 @@ function geoip_detect2_get_client_ip() {
 	if (!$ip)
 		$ip = '::1'; // By default, use localhost
 	
+	// @deprecated: this filter was added by mistake
 	$ip = apply_filters('geoip2_detect2_client_ip', $ip, $ip_list);
+	// this is the correct one!
+	$ip = apply_filters('geoip_detect2_client_ip', $ip, $ip_list);
 	
 	return $ip;
 }
@@ -225,7 +244,12 @@ function geoip_detect2_get_external_ip_adress($unfiltered = false) {
 
 	if (!$ip_cache) {
 		$ip_cache = _geoip_detect_get_external_ip_adress_without_cache();
-		set_transient('geoip_detect_external_ip', $ip_cache, GEOIP_DETECT_IP_CACHE_TIME);
+		
+		$expiryTime = GEOIP_DETECT_IP_CACHE_TIME;
+		if (empty($ip_cache) || $ip_cache === '0.0.0.0')
+			$expiryTime = GEOIP_DETECT_IP_EMPTY_CACHE_TIME;
+		
+		set_transient('geoip_detect_external_ip', $ip_cache, $expiryTime);
 	}
 	
 	$ip_cache = apply_filters('geoip_detect_get_external_ip_adress', $ip_cache);