+++ wp-content/plugins/geoip-detect/lib/geonames/data/country-names.php
@@ -1,5 +1,5 @@
 <?php
-// Generated at Sat, 03 Sep 2016 09:01:32 +0200 
+// Generated at Mon, 13 Mar 2017 12:14:10 +0100 
 return array (
   'en' => 
   array (